Prettify Webapp Output
Update webapp backend to version 2. The Python ML explanation library, ELI5, does a great job visualizing the prediction. However, it did not do as well when it comes to handling categorical outputs encoded as one-hot vectors. The upside is that it took less than ten lines of code changes to the model notebook to make the model compatible with ELI5.

As a side note, the `eli5.prediction object` from eli5.explain_prediction can be directly returned to Flask. No extra HTML tweaks required. ;)